camera: refactor (add qt_camera_base)
